+++ src/IO/ConsoleIO.php
@@ -21,7 +21,7 @@ use Symfony\Component\Console\Terminal;
 
 class ConsoleIO implements IOInterface
 {
-    const MAX_LINE_LENGTH = 120;
+    public const MAX_LINE_LENGTH = 120;
 
     /**
      * @var InputInterface
@@ -75,6 +75,19 @@ class ConsoleIO implements IOInterface
     public function isInteractive()
     {
         return $this->input->isInteractive();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function formatFileSize($bytes, $dec = 2)
+    {
+        $suffix = ['B', 'KiB', 'MiB', 'GiB', 'TiB', 'PiB', 'EiB', 'ZiB', 'YiB'];
+
+        $base = 1024;
+        $class = min((int) log($bytes, $base), count($suffix) - 1);
+
+        return sprintf("%1.{$dec}f", $bytes / pow($base, $class)) . ' ' . $suffix[$class];
     }
 
     /**
@@ -454,16 +467,16 @@ class ConsoleIO implements IOInterface
         $chars = substr(str_replace(PHP_EOL, "\n", $this->bufferedOutput->fetch()), -2);
 
         if (!isset($chars[0])) {
-            return $this->newLine(); //empty history, so we should start with a new line.
+            return $this->newLine(); // empty history, so we should start with a new line.
         }
-        //Prepend new line for each non LF chars (This means no blank line was output before)
+        // Prepend new line for each non LF chars (This means no blank line was output before)
         $this->newLine(2 - substr_count($chars, "\n"));
     }
 
     private function autoPrependText()
     {
         $fetched = $this->bufferedOutput->fetch();
-        //Prepend new line if last char isn't EOL:
+        // Prepend new line if last char isn't EOL:
         if ("\n" !== substr($fetched, -1)) {
             $this->newLine();
         }