+++ src/IO/IOInterface.php
@@ -12,6 +12,14 @@ interface IOInterface
     public function isInteractive();
 
     /**
+     * @param int $bytes
+     * @param int $dec
+     *
+     * @return string
+     */
+    public function formatFileSize($bytes, $dec = 2);
+
+    /**
      * Gets argument by name.
      *
      * @param string $name The name of the argument
@@ -60,9 +68,9 @@ interface IOInterface
      * @param string|array $question The question to ask
      * @param string $default The default answer if none is given by the user
      *
-     * @throws \RuntimeException If there is no data to read in the input stream
-     *
      * @return string The user answer
+     *
+     * @throws \RuntimeException If there is no data to read in the input stream
      */
     public function ask($question, $default = null);