+++ src/Package/PackageConstants.php
@@ -4,5 +4,5 @@ namespace Meteor\Package;
 
 final class PackageConstants
 {
-    const PATCH_DIR = 'to_patch';
+    public const PATCH_DIR = 'to_patch';
 }